Revert "New file structure and kmer query"
This reverts commit dc304359976a4acad85b50ed8f76f60cf046524e.
